requirements-test.txt
pytest
pytest-django
psycopg2-binary
psycopg
pytest-cov==4.1.0
